+++ zoomimage-core-coil2/src/main/kotlin/com/github/panpf/zoomimage/coil/CoilTileBitmapCache.kt
@@ -37,7 +37,6 @@ class CoilTileBitmapCache(private val imageLoader: ImageLoader) : TileBitmapCach
         tileBitmap: TileBitmap,
         imageUrl: String,
         imageInfo: ImageInfo,
-        disallowReuseBitmap: Boolean
     ): TileBitmap? {
         tileBitmap as AndroidTileBitmap
         val bitmap = tileBitmap.bitmap!!