+++ zoomimage-core-sketch/src/nonAndroidMain/kotlin/com/github/panpf/zoomimage/sketch/SketchTileBitmapCache.kt
@@ -19,6 +19,7 @@ package com.github.panpf.zoomimage.sketch
 import com.github.panpf.sketch.Sketch
 import com.github.panpf.sketch.asSketchImage
 import com.github.panpf.zoomimage.subsampling.BitmapFrom
+import com.github.panpf.zoomimage.subsampling.ImageInfo
 import com.github.panpf.zoomimage.subsampling.SkiaTileBitmap
 import com.github.panpf.zoomimage.subsampling.TileBitmap
 import com.github.panpf.zoomimage.subsampling.TileBitmapCache
@@ -39,8 +40,7 @@ actual class SketchTileBitmapCache actual constructor(
         key: String,
         tileBitmap: TileBitmap,
         imageUrl: String,
-        imageInfo: com.github.panpf.zoomimage.subsampling.ImageInfo,
-        disallowReuseBitmap: Boolean
+        imageInfo: ImageInfo,
     ): TileBitmap? {
         tileBitmap as SkiaTileBitmap
         val bitmap = tileBitmap.bitmap